refactor(app): remove redundant WifiStatsManager and consume shared models and formatting from :core
