+++ app/src/main/java/com/vinnovateit/latch/common/util/StatsUtils.kt
@@ -1,7 +1,9 @@
 package com.vinnovateit.latch.common.util
 
 import androidx.compose.ui.graphics.Path
-import com.vinnovateit.latch.domain.model.LiveDataPoint
+import com.vinnovateit.latch.core.model.LiveDataPoint
+import com.vinnovateit.latch.core.stats.formatBitsPerSecond
+import com.vinnovateit.latch.core.stats.formatBytes
 import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
@@ -39,30 +41,11 @@ data class GraphData(
     }
 }
 
-fun formatBytes(bytes: Long, unit: String = "B/s"): Pair<String, String> = when {
-    unit == "bps" -> formatBitsPerSecond(bytes)
-    bytes < 1_024L -> bytes.toString() to "B"
-    bytes < 1_048_576L -> "%.1f".format(bytes / 1_024f) to "KB"
-    bytes < 1_073_741_824L -> "%.1f".format(bytes / 1_048_576f) to "MB"
-    else -> "%.2f".format(bytes / 1_073_741_824f) to "GB"
-}
+fun formatBytes(bytes: Long, unit: String = "B/s"): Pair<String, String> =
+    com.vinnovateit.latch.core.stats.formatBytes(bytes, unit)
 
-fun formatBitsPerSecond(bytesPerSecond: Long, unit: String = "bps"): Pair<String, String> {
-    if (unit == "B/s") {
-        val (value, byteUnit) = formatBytes(bytesPerSecond, "B/s")
-        return value to "$byteUnit/s"
-    }
-
-    val bitsPerSecond = bytesPerSecond * 8
-    val (value, bitUnit) = when {
-        bitsPerSecond < 1_000L -> bitsPerSecond.toString() to "bps"
-        bitsPerSecond < 1_000_000L -> "%.1f".format(bitsPerSecond / 1_000f) to "Kbps"
-        bitsPerSecond < 1_000_000_000L -> "%.1f".format(bitsPerSecond / 1_000_000f) to "Mbps"
-        else -> "%.2f".format(bitsPerSecond / 1_000_000_000f) to "Gbps"
-    }
-    return value to bitUnit
-}
-
+fun formatBitsPerSecond(bytesPerSecond: Long, unit: String = "bps"): Pair<String, String> =
+    com.vinnovateit.latch.core.stats.formatBitsPerSecond(bytesPerSecond, unit)
 
 fun formatDurationDynamic(ms: Long): String {
     if (ms < 0) return "0s"
