+++ app/src/main/java/com/vinnovateit/latch/features/home/HomeScreen.kt
@@ -64,8 +64,8 @@ import com.vinnovateit.latch.features.wifi.detector.WiFiStateDetector
 import com.vinnovateit.latch.R
 import com.vinnovateit.latch.common.ui.LeafOverlay
 import com.vinnovateit.latch.common.util.TooltipHint
-import com.vinnovateit.latch.domain.model.LiveDataPoint
-import com.vinnovateit.latch.domain.model.SessionSummary
+import com.vinnovateit.latch.core.model.LiveDataPoint
+import com.vinnovateit.latch.core.model.SessionSummary
 import com.vinnovateit.latch.features.home.components.SpectrumCard
 import com.vinnovateit.latch.features.settings.manager.SettingsManager
 import com.vinnovateit.latch.features.wifi.background.ForegroundService
